PROG_GameMode_NewFeatures/Changes_GameStateAndTeamFunctionality
Added GameMode Folder for us to put our code into.
Added TeamIdentifier Enum which is used by the CTFPlayerState and Team classes.
Added CTFGameState and CTFPlayerState.
GameState holds teams, assigns players to those teams and binds to event in PlayerState.
PlayerState holds team identifier, points earned per player and other info that may or may not be used.
PlayerState is used to assign points to whichever team the player is on.
CapturePoint 's AddPoints function calls the CTFPlayerState's addScore function.
Flag has a pointValue variable so we can decide how many points a flag is worth and add it to the team that captured it.
CTFGameMode has some placeholder functions (PostLogin and logout) for when the networking team gets a hold of it.