PROG_GameMode_New_Feature_Flag_Indicator
Added a flag indicator that shows where the flag is on player screen. If the player turns so the flag is not in their current line of site. The indicator will move to the side of the screen closest to it. Known issue: only works on server side and indicator is also on the back of flag holders.