PROG_GameMode_Change_Flag Indicators
Flags indicator changes based on which team is carrying it. Once we have an icon for neutral flags then the icon for when the flag is neutral can be changed to that (currently blue crystal icon). That is found at the end of the double false track.

Added the mainflag to the flagindicators.